Complete operation logic and display updates
Added event handler to equal sign that calls the operate() function.
The logic is to split the displayValue string, using a regex, into a
temporary array with two items. The items are then converted into
numbers and set as firstNum and secondNum values.
Modified the operate() function so that it rounds the result to at most
two decimal places and it now displays an error if trying to divide by
zero.
Updated html.index and css.style for consistency and better readability.