Refactor variable acquisition and buttons handlers
Added emptyDisplay() function that clears the display once a number is
entered after an operator or if the leading character is a zero.
Added the evaluate() function which sets the current display content to
the secondNum variable.
Modified the code so that evaluate() can be called by pressing the equal
sign or with two consecutive operators.