Edit style, solve divide by zero issue
Added images directory and edited index.html.
Reverted to use display flex instead of grid.
Solved the divide by zero issue inside evaluate() function, now if the
secondNum is zero the calculator will display an error on display and
set shouldEmptyDisplay to true so that a new number can be input.